requirements.txt
altair==4.2.0
hiplot==0.1.33
matplotlib==3.5.3
numpy==1.23.2
pandas==1.4.4
Pillow==9.2.0
plotly==5.10.0
seaborn==0.11.2
streamlit==1.12.2
